Finished top 10 request
